// src/types/styles/border.ts
export type BorderTypes = {
  border?: string;
  border_top?: string;
  border_right?: string;
  border_bottom?: string;
  border_left?: string;
  border_radius?: string;
  border_width?: string;
  border_left_width?: string;
  border_right_width?: string;
  border_top_width?: string;
  border_bottom_width?: string;
  border_color?: string;
  border_left_color?: string;
  border_right_color?: string;
  border_top_color?: string;
  border_bottom_color?: string;
  border_top_left_radius?: string;
  border_top_right_radius?: string;
  border_bottom_left_radius?: string;
  border_bottom_right_radius?: string;
  border_style?: string;
  border_left_style?: string;
  border_top_style?: string;
  border_right_style?: string;
  border_bottom_style?: string;
  outline?: string;
  outline_style?: string;
  outline_color?: string;
  outline_width?: string;
  outline_offset?: string;
};