// src/types/styles/scroll.ts
export type ScrollTypes = {
  scrollbar_width?: string;
  scrollbar_color?: string;
  overflow_block?: string;
  overflow_inline?: string;
  overflow_x?: string;
  overflow_y?: string;
  overflow?: string;
  overflow_clip_margin?: string;
  scrollbar_gutter?: string;
  scroll_behavior?: string;
  scroll_margin?: string;
  scroll_padding?: string;
  scroll_snap_align?: string;
  scroll_snap_stop?: string;
  scroll_snap_type?: string;
  webkit_scrollbar?: string; // ::-webkit-scrollbar pseudo-element
  scroll_container?: string; // scroll container glossary term
  scrollbar_aria_role?: string; // scrollbar ARIA role
};